hide secp256k1
