--- api/src/primitives/mod.rs
@@ -1,2 +1,0 @@
-/// Module for asymmetric encryption.
-pub mod asymmetric_encryption;
